Fix parsing 'client_oid' in websocket feed (#4)
* Fix parsing 'client_oid' in websocket feed

This field is showing up as client_oid:"" but the empty string cannot
be parsed as a UUID and was causing the parse to fail. We are 
ignoring this field if it cannot be parsed as a UUID now. It has become
"opaque" in that case.

Using Aeson's default-value combinator and clarify comment on when
parsing results in Nothing default